requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# inference.py: 6
# rem/Utils/utils.py: 3
Pillow_SIMD == 4.1.1.post0

# rem/Criterions/CrossEntropy.py: 1
# rem/Criterions/L1Loss.py: 1
# rem/Criterions/L2Loss.py: 1
# rem/Layers/Averagepool.py: 1
# rem/Layers/BN.py: 1
# rem/Layers/Conv2d.py: 1
# rem/Layers/Dropout.py: 3
# rem/Layers/Layer.py: 1
# rem/Layers/Linear.py: 1
# rem/Layers/Maxpool.py: 1
# rem/Layers/NonLinear.py: 1
# rem/Layers/SpatialBN.py: 1
# rem/Models/Model.py: 4
# rem/Optimizers/Optimizer.py: 3
# rem/Utils/utils.py: 1
# test.py: 2
# tester.py: 1
# train.py: 8
numpy == 1.13.1
